Equivalence Tests Draft 2: Documentation, 131kb report, checking for matching types in equivalence tests,simplify data contract through JAX model, handle casts that weren't caught the first time, cast LossOutput to output_dtype

+++ src/alphagenome_pt/normalize.py
@@ -129,7 +129,7 @@ def _normalize_splice_site_positions(
 
     splice_site_positions = splice_site_positions.to(
         device=dna_onehot.device,
-        dtype=torch.long,
+        dtype=torch.int32,
     )
     invalid_position = (
         (splice_site_positions < -1) | (splice_site_positions >= seq_len)
